# planning-poker/internal/infra/boundaries/http/swagger/swagger.yaml
basePath: /
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  http.GetAllRoomsStateClient:
    properties:
      ID:
        type: string
      IsOwner:
        type: boolean
      IsSpectator:
        type: boolean
      Name:
        type: string
    type: object
  http.GetAllRoomsStateResponse:
    properties:
      Clients:
        items:
          $ref: '#/definitions/http.GetAllRoomsStateClient'
        type: array
      ID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for the Planning Poker application. This API allows creating rooms,
    managing clients, voting, and administration.
  title: Planning Poker API
  version: 1.0.0
paths:
  /admin/rooms:
    get:
      description: Returns the state of all rooms (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.GetAllRoomsStateResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all rooms state
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer token format. Use: Bearer <api_key>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"